requirements.txt
openai>=1.0.0
chromadb>=0.4.0
pandas>=2.0.0
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
